requirements.txt
customtkinter
feedparser
pandas
beautifulsoup4
joblib
requests
tkhtmlview
Pillow
